Opening and constructing
Got testing set up a little more.

Wrote tests and implemented the constructor.
The constructor now only takes stream options and inits the object.

Wrote tests and implemented the the `open` instance function.

Wrote tests and implementd the `open` and `openPort` module functions.
This is the recommended way to get an open SerialPort instance by passing options and an optional callback.

Opening a port now gets `comName` from the options object.
Opening callback is now subscribed to the open event.