feat: add optional end event for piping (#1926)
This will allow streams to end. We haven't done this because it allows a stream to be reopened and data to be continued to be read. 

Adds `endOnClose` to stream options so the stream will emit the end event when the port is closed. This will signal any piped parsers to end as well. If the port is reopened the parsers won’t function.