Upgrade to Lerna v7 and drop support for Node 14 (#2686)
There were several dependencies which couldn't be updated until node 14 was dropped, additionally Lerna 7 required a few changes, so I've bundled these together to avoid breaking interdependencies. Hopefully this can be the basis for the next release.

This includes:

    Dropping Node 14 (githhub action change, @tsconfig/node18, etc)
    Switch to NPM workspaces
    Upgrade to Lerna v7
    Upgrade to Prettier v3
    Upgrade to bindings-cpp v12
    Upgrade to commander v11
    Upgrade to enquirer v2.4.1

Please note the switch to NPM workspaces results in a single package-lock.json in the root. The individual lock files are no longer used, but I have not deleted these yet.